Update AccessibilityAssessmentTools.md
One of the links had an error that meant the whole code was showing up rather than just the text with the link. It was "An example of a Mozilla browser plugin ....". I added a space after the URL in the code and now it shows up correctly on output.